フロントエンドのディレクトリ構造整理
